requirements.txt
pyjwt[crypto]
sqlmodel
fastapi
mongoengine
uvicorn[standard]